+++ UnityEngine.TestRunner/TestRunner/RuntimeTestRunnerFilter.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Text.RegularExpressions;
 using NUnit.Framework.Interfaces;
 using NUnit.Framework.Internal;
 using NUnit.Framework.Internal.Filters;
@@ -22,7 +23,7 @@ namespace UnityEngine.TestTools.TestRunner.GUI
             var filters = new List<ITestFilter>();
 
             AddFilters(filters, testNames, (s) => new FullNameFilter(s));
-            AddFilters(filters, groupNames, (s) => new FullNameFilter(s) {IsRegex = true});
+            AddFilters(filters, groupNames, OptimizedGroupFilter);
             AddFilters(filters, assemblyNames, (s) => new AssemblyNameFilter(s));
             AddFilters(filters, categoryNames, (s) => new CategoryFilterExtended(s) {IsRegex = true});
 
@@ -34,6 +35,28 @@ namespace UnityEngine.TestTools.TestRunner.GUI
             return filters.Count == 0 ? TestFilter.Empty : new AndFilter(filters.ToArray());
         }
 
+        static FullNameFilter OptimizedGroupFilter(string s)
+        {
+            if (s.Length >= 2)
+            {
+                // A common case is that the regex we are filtering by is of the form
+                //   ^JUST_A_NAME$
+                // so we have a regex that matches _precisely_ one string. This can be done without a regex, which
+                // is much much faster.
+                if (s[0] == '^' && s[s.Length - 1] == '$')
+                {
+                    var raw = s.Substring(1, s.Length - 2);
+                    var escaped = Regex.Escape(raw);
+                    // if a regex is the same when we escape it, it means that it doesn't contain any characters
+                    // with any meaning in the regex. Hence the regex is just a plain name.
+                    if (raw.Equals(escaped, StringComparison.Ordinal))
+                        return new FullNameFilter(raw);
+                }
+            }
+
+            return new FullNameFilter(s) { IsRegex = true };
+        }
+        
         private static void AddFilters(List<ITestFilter> filters, string[] values, Func<string, TestFilter> builder)
         {
             if (values == null || values.Length == 0)